Add Dropbox basic support and shift Amazon to S3 only
Added basic support for filename searching in Dropbox. Hash search is
not supported in their API at this time.
Shifted Amazon service to be S3 due to differences in the different AWS
service structures. When additional services are configured we should
look into shifting to an aws_core which handles creation of the clients
and then the interface classes.
Updated the base interface initializer details to allow a python logging
interface to be passed through. May want to investigate a better
mechanism for this type of operation but for now it works.